% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ModCM}
\alias{ModCM}
\title{Modyfikacja testu Cramera-von Misesa}
\usage{
ModCM(tab)
}
\arguments{
\item{tab}{niepusty, bez braków danych wektor z danymi.}
}
\description{
Modyfikacja testu Cramera-von Misesa
}

